Acorn: Started on very basic unit tests
We use gtest to test our Application. For this an enable_test option was added.
For now profiling and tests are mutually exclusive, but
in the future I want to add macros that conditionally disable
profiling on testing automatically.

There also is much more work to be done in logging (Either a new
AC_*_TRACE/DEBUG/etc Macro or a good spdlog setup.
And of course writing A LOT more tests.

For now we only do basic tests for the LayerStack, since that was
the easiest class to start testing ;)