IDLParser: Switched to inja and started implementation
We will use inja for templating our c++ files.
I also started writing some wrappers. Currently
the enum wrapper works and there is some basic
implementation logic for the interface wrapper.

I also added a script to convert template (.tpl) files
written in the inja templating language to a header file,
because I want to have the template as strings
in the executable (for now)
so we don't have to worry about path resolution etc.
This will probaby change in the future, since this is not a great system.
It's likely going to end up as a method in the SourceGenerator,
so we can immediately parse the files into inja::Template objects.

Currently I'm working on getting a basic wrapper implementation up and running,
therefore commits will not be very clean.